requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-multipart
passlib
python-jose
starlette
pydantic
asgiref